Began on HTTP interface

+++ lib/Interlock/src/Types.hpp
@@ -1,4 +1,4 @@
 #pragma once
 #include <stdint.h>
 
-typedef uint32_t RFIDCard;
+typedef uint16_t RFIDCard;
